// web/src/generated/api/models/chatMessageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageMetadata } from './chatMessageMetadata';
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessageResponse {
  id: string;
  sessionId: string;
  /** @nullable */
  clientMessageId: string | null;
  /** @nullable */
  replyToMessageId: string | null;
  role: ChatMessageRole;
  content: string;
  structuredData: ChatMessageMetadata | null;
  createdAt: string;
}