// web/src/generated/api/models/chatTurnResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageResponse } from './chatMessageResponse';
import type { ChatSessionResponse } from './chatSessionResponse';

export interface ChatTurnResponse {
  session: ChatSessionResponse;
  userMessage: ChatMessageResponse;
  assistantMessage: ChatMessageResponse;
}