// web/src/generated/api/models/confirmChatBookingResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentResponse } from './appointmentResponse';
import type { ChatMessageResponse } from './chatMessageResponse';
import type { ChatSessionResponse } from './chatSessionResponse';

export interface ConfirmChatBookingResponse {
  session: ChatSessionResponse;
  assistantMessage: ChatMessageResponse;
  appointment: AppointmentResponse;
}